requirements.txt
# Python dependencies for News Aggregator Project

# Ingestion & Processing
requests
beautifulsoup4
feedparser
aiohttp

# transformer
torch

# Summarization & Clustering
numpy
scikit-learn
google-generativeai
yake
nltk

# Backend & API
Flask
flask-socketio
python-dotenv
flask-cors==4.0.0
gunicorn==21.2.0
pydantic
tenacity
loguru
PyJWT==2.8.0

# Caching & Storage
redis
supabase>=2.0.0

# Testing
pytest

flask_cors
flask_restx

# Scheduling & Background Tasks
schedule